+++ modules/server/router.js
@@ -1,11 +1,11 @@
-
-var gracenode = require('../../../');
+var gracenode = require('../../');
 var log = gracenode.log.create('server-router');
 
 var EventEmitter = require('events').EventEmitter;
 
 var config = null;
 var rerouteMap = {};
+var controllerMap = {};
 
 module.exports = new EventEmitter();
 
@@ -21,6 +21,25 @@ module.exports.readConfig = function (configIn) {
 		log.verbose('rerouting mapped:', rerouteMap);
 	
 	}
+};
+
+module.exports.setup = function (cb) {
+	gracenode.lib.walkDir(config.controllerPath, function (error, list) {
+		if (error) {
+			return cb(error);
+		}
+		for (var i = 0, len = list.length; i < len; i++) {
+			var sep = list[i].file.replace(config.controllerPath, '').split('/');
+			var controller = sep[0];
+			var method = sep[1].substring(0, sep[1].lastIndexOf('.'));
+			if (!controllerMap[controller]) {
+				controllerMap[controller] = {};
+			}
+			controllerMap[controller][method] = true;
+		}
+		log.verbose('controllers and methods mapped:', controllerMap);
+		cb();
+	});
 };
 
 module.exports.handle = function (url, res) {
@@ -65,12 +84,28 @@ function parseUrl(url) {
 			return item;
 		}
 	});
+
+	var controller = parsed[0] || null;
 	// if there is no method in URL, gracenode will look for index.js
+	var method = parsed[1] || 'index';
+
+	var notFound = null;
+	
+	// check the controller map
+	if (!controllerMap[controller]) {
+		notFound = new Error('controller ' + controller + ' not found');
+	} else if (!controllerMap[controller][method]) {
+		notFound = new Error('controller method ' + controller + '/' + method + ' not found');
+	}
+
+	log.verbose('controller and method found: ', controller + '/' + method);
+	
 	return {
-		controller: parsed[0] || null,
-		method: parsed[1] || 'index',
+		controller: controller,
+		method: method,
 		parameters: parsed.length > 2 ? parsed.splice(2) : [],
-		originalRequest: null
+		originalRequest: null,
+		notFound: notFound
 	};
 }